Refactor file structure.
Move things around to separate things a bit.
